Added support for single quoted character literals
This also fixes a bug where the following:

if test = '"' then

would be interpreted as a double quoted string beginning when " is hit
spanning the string multiple lines until the next " is hit.

Note that a single quoted " character surrounded by additional characters
that extends the string beyond length 1 will span an incorrect match.

This is intended behaviour as '' may only quote a single character.