More consistent coloring and naming
Procedure, package and function keywords are now marked
as keyword.other.ada instead of their 'specific' names.

This makes the syntax coloring more consistent with the rest
of the code.

Package names are now marked as entity.other.attribute-name as the
previously used naming was supported only by a fraction of
existing themes.